requirements.txt
pandas
scikit-learn
numpy
seaborn
ipykernel
Flask
#this is more stable version
mlflow==2.2.2
dvc
apache-airflow
azure-storage-blob


-e .